Refactor folder structure Add start script
