+++ api/config/config.js
@@ -1,5 +1,6 @@
 const config = {
-  db:`mongodb://localhost:27017/keep-api`
+  db:"mongodb://localhost:27017/keep-api",
+  port:process.env.port ||9090
 }
 
 module.exports = config;
