+++ package.json
@@ -4,7 +4,8 @@
   "description": "This is the official backend for keep-api",
   "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start":"node --experimental-modules app.mjs"
   },
   "repository": {
     "type": "git",
@@ -23,6 +24,7 @@
   "homepage": "https://github.com/ebenKB/keep-api#readme",
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "dotenv": "^6.1.0",
     "express": "^4.16.4",
     "jsonwebtoken": "^8.3.0",
     "mongoose": "^5.3.5"
